Refactor chat input context handling and improve component structure: replace direct context usage with a safe context hook in multiple chat components to enhance error handling. Move feature color and icon definitions to a dedicated constants file for better organization. Update ESLint configuration to ignore 'src-tauri' directory. Introduce new hooks for managing Tauri secrets and Pinecone configuration, improving code modularity and maintainability.
