+++ frontend/src/components/ui/TauriPineconeSetupModal.tsx
@@ -272,37 +272,3 @@ export function TauriPineconeSetupModal({ isOpen, onClose, onSaveSuccess }: Taur
   );
 }
 
-// Hook to check if Pinecone is configured
-export function usePineconeConfigured() {
-  const [isConfigured, setIsConfigured] = useState<boolean | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-
-  const checkConfiguration = useCallback(async () => {
-    if (!isTauri()) {
-      setIsConfigured(true); // Assume configured in web mode
-      setIsLoading(false);
-      return;
-    }
-
-    try {
-      setIsLoading(true);
-      const secrets = await getSecrets();
-      const configured = !!(
-        secrets.pinecone_api_key?.trim() &&
-        secrets.pinecone_index_name?.trim()
-      );
-      setIsConfigured(configured);
-    } catch {
-      setIsConfigured(false);
-    } finally {
-      setIsLoading(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    checkConfiguration();
-  }, [checkConfiguration]);
-
-  return { isConfigured, isLoading, refetch: checkConfiguration };
-}
-
