+++ frontend/src/features/chat/components/input/ChatDragOverlay.tsx
@@ -5,7 +5,7 @@
  * Can be used standalone or with ChatInputContext
  */
 
-import { useChatInputContext } from './ChatInputContext';
+import { useChatInputContextSafe } from './ChatInputContext';
 
 export interface ChatDragOverlayProps {
   /** Whether to show the overlay (optional if using context) */
@@ -13,13 +13,8 @@ export interface ChatDragOverlayProps {
 }
 
 export function ChatDragOverlay({ visible: propVisible }: ChatDragOverlayProps = {}) {
-  // Try to use context, but fall back to props
-  let contextValue: ReturnType<typeof useChatInputContext> | null = null;
-  try {
-    contextValue = useChatInputContext();
-  } catch {
-    // Not in a ChatInput context, use props
-  }
+  // Use safe context hook - returns null if not in ChatInput context
+  const contextValue = useChatInputContextSafe();
 
   const visible = propVisible ?? contextValue?.isDragging ?? false;
