+++ frontend/src/features/chat/components/input/ChatInputMetrics.tsx
@@ -5,7 +5,7 @@
  * Can be used standalone with props or with ChatInputContext
  */
 
-import { useChatInputContext } from './ChatInputContext';
+import { useChatInputContextSafe } from './ChatInputContext';
 
 export interface ChatInputMetricsProps {
   /** Character count (optional if using context) */
@@ -24,13 +24,8 @@ export function ChatInputMetrics({
   tokenCount: propTokenCount,
   attachedFileCount: propAttachedFileCount,
 }: ChatInputMetricsProps) {
-  // Try to use context, but fall back to props
-  let contextValue: ReturnType<typeof useChatInputContext> | null = null;
-  try {
-    contextValue = useChatInputContext();
-  } catch {
-    // Not in a ChatInput context, use props
-  }
+  // Use safe context hook - returns null if not in ChatInput context
+  const contextValue = useChatInputContextSafe();
 
   const charCount = propCharCount ?? contextValue?.charCount ?? 0;
   const wordCount = propWordCount ?? contextValue?.wordCount ?? 0;
