+++ frontend/src/pages/settings/AISettings.tsx
@@ -4,8 +4,10 @@ import { useSettingsStore } from '../../store/settings-store';
 import { useAuthStore } from '../../store/auth-store';
 import { useOllamaDownloadStore, formatBytes, formatSpeed, formatTimeRemaining } from '../../store/ollama-download-store';
 import { Modal } from '../../components/ui/Modal';
-import { TauriProviderApiKeyInput, useTauriSecrets } from '../../components/ui/TauriApiKeysManager';
-import { TauriPineconeSetupModal, usePineconeConfigured } from '../../components/ui/TauriPineconeSetupModal';
+import { TauriProviderApiKeyInput } from '../../components/ui/TauriApiKeysManager';
+import { useTauriSecrets } from '../../components/ui/use-tauri-secrets';
+import { TauriPineconeSetupModal } from '../../components/ui/TauriPineconeSetupModal';
+import { usePineconeConfigured } from '../../components/ui/use-pinecone-configured';
 import { useAIHealth } from '../../features/ai/hooks/use-ai-health';
 import { formatModelName } from '../../utils/model-name-formatter';
 import { toast } from '../../hooks/use-toast';
